paths: moved CollectPathsToFormat to new package
I'm trying to break apart the `util.go` file into other relevant
packages since I hate util dumping grounds.
