+++ engine/util.go
@@ -14,60 +14,7 @@
 
 package engine
 
-import (
-	"fmt"
-	"strings"
-
-	"github.com/bmatcuk/doublestar/v4"
-	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
-)
-
-func CollectPathsToFormat(include, exclude []string) ([]string, error) {
-	includedPaths := []string{}
-	for _, pattern := range include {
-		globMatches, err := doublestar.FilepathGlob(pattern)
-		if err != nil {
-			return nil, err
-		}
-		includedPaths = append(includedPaths, globMatches...)
-	}
-
-	pathsToFormatSet := map[string]struct{}{}
-	for _, path := range includedPaths {
-		if len(exclude) == 0 {
-			pathsToFormatSet[path] = struct{}{}
-			continue
-		}
-		excluded := false
-		for _, pattern := range exclude {
-			match, err := doublestar.Match(pattern, path)
-			if err != nil {
-				return nil, err
-			}
-			if match {
-				excluded = true
-			}
-		}
-		if !excluded {
-			pathsToFormatSet[path] = struct{}{}
-		}
-	}
-	pathsToFormat := []string{}
-	for path := range pathsToFormatSet {
-		pathsToFormat = append(pathsToFormat, path)
-	}
-	return pathsToFormat, nil
-}
-
-func MultilineStringDiff(a, b string) string {
-	return cmp.Diff(
-		a, b,
-		cmpopts.AcyclicTransformer("multiline", func(s string) []string {
-			return strings.Split(s, "\n")
-		}),
-	)
-}
+import "fmt"
 
 type DryRunDiffs struct {
 	diffs map[string]string
